+++ dali/kernels/CMakeLists.txt
@@ -14,9 +14,9 @@
 
 project(dali_kernel)
 
-add_subdirectory(audio)
 add_subdirectory(common)
 add_subdirectory(imgproc)
+add_subdirectory(signal)
 add_subdirectory(slice)
 add_subdirectory(test)
